Feature: 12 confirm block hashes - Total mess
