+++ orchestrator/consensus/service.go
@@ -3,10 +3,8 @@ package consensus
 import (
 	"context"
 	"fmt"
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/lukso-network/lukso-orchestrator/orchestrator/db"
 	"github.com/lukso-network/lukso-orchestrator/orchestrator/db/iface"
-	"github.com/lukso-network/lukso-orchestrator/orchestrator/db/kv"
 	"github.com/lukso-network/lukso-orchestrator/orchestrator/rpc/api/events"
 	"github.com/lukso-network/lukso-orchestrator/orchestrator/utils"
 	"github.com/lukso-network/lukso-orchestrator/shared"
@@ -44,9 +42,10 @@ type Service struct {
 	canonicalizeChan            chan uint64
 	isWorking                   bool
 	canonicalizeLock            *sync.Mutex
-	invalidationWorkPayloadChan chan invalidationWorkPayload
+	invalidationWorkPayloadChan chan *invalidationWorkPayload
 	errChan                     chan databaseErrors
 	started                     bool
+	ctx                         context.Context
 }
 
 // Start service should be registered only after Pandora and Vanguard notified about:
@@ -93,7 +92,7 @@ func New(
 ) (service *Service) {
 	stopChan := make(chan bool)
 	canonicalizeChain := make(chan uint64)
-	invalidationWorkPayloadChan := make(chan invalidationWorkPayload, 10000)
+	invalidationWorkPayloadChan := make(chan *invalidationWorkPayload, 10000)
 	errChan := make(chan databaseErrors, 10000)
 
 	return &Service{
@@ -110,6 +109,7 @@ func New(
 		invalidationWorkPayloadChan: invalidationWorkPayloadChan,
 		// errChan is internal so it is created on the fly
 		errChan: errChan,
+		ctx:     ctx,
 	}
 }
 
@@ -124,7 +124,7 @@ func invokeInvalidation(
 	batchLimit uint64,
 	skipChan chan bool,
 	databaseErrorsChan chan databaseErrors,
-	invalidationWorkPayloadChan chan invalidationWorkPayload,
+	invalidationWorkPayloadChan chan *invalidationWorkPayload,
 ) {
 	possibleSkippedPair := make([]*events.RealmPair, 0)
 	latestSavedVerifiedRealmSlot := realmDB.LatestVerifiedRealmSlot()
@@ -177,7 +177,7 @@ func invokeInvalidation(
 	log.WithField("pandoraRange", pandoraRange).WithField("vanguardRange", vanguardRange).
 		Trace("Invalidation with range of blocks")
 
-	invalidationWorkPayloadChan <- invalidationWorkPayload{
+	invalidationWorkPayloadChan <- &invalidationWorkPayload{
 		invalidationStartRealmSlot: invalidationStartRealmSlot,
 		fromSlot:                   fromSlot,
 		possibleSkippedPair:        possibleSkippedPair,
@@ -306,7 +306,7 @@ func (service *Service) workLoop() {
 
 		latestVerifiedRealmSlot := realmDB.LatestVerifiedRealmSlot()
 
-		// This is naive, but might workHeader
+		// This is naive, but might work
 		// We need to have at least one pair to start invalidation.
 		// It might lead to 2 pairs on one side, or invalidation stall,
 		// But ATM I do not have quicker and better idea
@@ -375,327 +375,6 @@ func (service *Service) workLoop() {
 		mergedHeadersChanBridge,
 		mergedChannelHandler,
 	)
-}
-
-// handlePreparedWork should be synchronous approach
-func handlePreparedWork(
-	vanguardHashDB db.VanguardHeaderHashDB,
-	pandoraHeaderHashDB db.PandoraHeaderHashDB,
-	realmDB db.RealmDB,
-	invalidationWorkPayload invalidationWorkPayload,
-	errChan chan databaseErrors,
-) {
-	vanguardBlockHashes := invalidationWorkPayload.vanguardHashes
-	pandoraHeaderHashes := invalidationWorkPayload.pandoraHashes
-	invalidationStartRealmSlot := invalidationWorkPayload.invalidationStartRealmSlot
-	fromSlot := invalidationWorkPayload.fromSlot
-	possibleSkippedPair := invalidationWorkPayload.possibleSkippedPair
-	vanguardOrphans := invalidationWorkPayload.vanguardOrphans
-	pandoraOrphans := invalidationWorkPayload.pandoraOrphans
-
-	var (
-		vanguardErr error
-		pandoraErr  error
-		realmErr    error
-	)
-
-	// TODO: move it to memory, and save in batch
-	// This is quite naive, but should work
-	for index, vanguardBlockHash := range vanguardBlockHashes {
-		slotToCheck := fromSlot + uint64(index)
-
-		if len(pandoraHeaderHashes) <= index {
-			break
-		}
-
-		pandoraHeaderHash := pandoraHeaderHashes[index]
-
-		// Potentially skipped slot
-		if nil == pandoraHeaderHash && nil == vanguardBlockHash {
-			possibleSkippedPair = append(possibleSkippedPair, &events.RealmPair{
-				Slot:          slotToCheck,
-				VanguardHash:  nil,
-				PandoraHashes: nil,
-			})
-
-			continue
-		}
-
-		// I dont know yet, if it is true.
-		// In my opinion INVALID state is 100% accurate only with blockShard verification approach
-		// TODO: add additional Sharding info check VanguardBlock -> PandoraHeaderHash when implementation on vanguard side will be ready
-		if nil == pandoraHeaderHash {
-			vanguardHeaderHash := &types.HeaderHash{
-				HeaderHash: vanguardBlockHash.HeaderHash,
-				Status:     types.Pending,
-			}
-			vanguardOrphans[slotToCheck] = vanguardHeaderHash
-
-			continue
-		}
-
-		if nil == vanguardBlockHash {
-			currentPandoraHeaderHash := &types.HeaderHash{
-				HeaderHash: pandoraHeaderHash.HeaderHash,
-				Status:     types.Pending,
-			}
-			pandoraOrphans[slotToCheck] = currentPandoraHeaderHash
-
-			continue
-		}
-
-		log.WithField("slot", slotToCheck).
-			WithField("hash", vanguardBlockHash.HeaderHash.String()).
-			Debug("I am inserting verified vanguardBlockHash")
-
-		vanguardErr = vanguardHashDB.SaveVanguardHeaderHash(slotToCheck, &types.HeaderHash{
-			HeaderHash: vanguardBlockHash.HeaderHash,
-			Status:     types.Verified,
-		})
-
-		log.WithField("slot", slotToCheck).
-			WithField("hash", pandoraHeaderHash.HeaderHash.String()).
-			Debug("I am inserting verified pandoraHeaderHash")
-
-		pandoraErr = pandoraHeaderHashDB.SavePandoraHeaderHash(slotToCheck, &types.HeaderHash{
-			HeaderHash: pandoraHeaderHash.HeaderHash,
-			Status:     types.Verified,
-		})
-
-		if nil != vanguardErr || nil != pandoraErr {
-			break
-		}
-
-		realmErr = realmDB.SaveLatestVerifiedRealmSlot(slotToCheck)
-	}
-
-	if nil != vanguardErr || nil != pandoraErr || nil != realmErr {
-		log.WithField("vanguardErr", vanguardErr).
-			WithField("pandoraErr", pandoraErr).
-			WithField("realmErr", realmErr).
-			Error("Got error during invalidation of pending queue")
-		errChan <- databaseErrors{
-			vanguardErr: vanguardErr,
-			pandoraErr:  pandoraErr,
-			realmErr:    realmErr,
-		}
-
-		return
-	}
-
-	// Resolve state of possible invalid pairs
-	latestSavedVerifiedRealmSlot := realmDB.LatestVerifiedRealmSlot()
-	log.WithField("possibleInvalidPairs", len(possibleSkippedPair)).
-		WithField("latestVerifiedRealmSlot", latestSavedVerifiedRealmSlot).
-		WithField("invalidationStartRealmSlot", invalidationStartRealmSlot).
-		Info("Requeue possible invalid pairs")
-
-	invalidationRange := latestSavedVerifiedRealmSlot - invalidationStartRealmSlot
-
-	// All of orphans and possibleSkipped are still pending
-	if 0 == invalidationRange {
-		log.WithField("invalidationStartRealmSlot", invalidationStartRealmSlot).
-			WithField("latestVerifiedRealmSlot", latestSavedVerifiedRealmSlot).
-			Warn("I did not progress any slot")
-
-		return
-	}
-
-	if invalidationRange < 0 {
-		log.Fatal("Got wrong invalidation range. This is a fatal bug that should never happen.")
-
-		return
-	}
-
-	// Mark all pandora Orphans as skipped
-	for slot := range pandoraOrphans {
-		pandoraErr = pandoraHeaderHashDB.SavePandoraHeaderHash(slot, &types.HeaderHash{
-			HeaderHash: common.Hash{},
-			Status:     types.Skipped,
-		})
-	}
-
-	// Mark all vanguard orphans as skipped
-	for slot := range vanguardOrphans {
-		vanguardErr = vanguardHashDB.SaveVanguardHeaderHash(slot, &types.HeaderHash{
-			HeaderHash: common.Hash{},
-			Status:     types.Skipped,
-		})
-	}
-
-	if nil != vanguardErr || nil != pandoraErr {
-		log.WithField("vanguardErr", vanguardErr).
-			WithField("pandoraErr", pandoraErr).
-			Error("Got error during invalidation of possible orphans")
-		errChan <- databaseErrors{
-			vanguardErr: vanguardErr,
-			pandoraErr:  pandoraErr,
-			realmErr:    realmErr,
-		}
-
-		return
-	}
-
-	pendingPairs := make([]*events.RealmPair, 0)
-
-	for _, pair := range possibleSkippedPair {
-		if nil == pair {
-			continue
-		}
-
-		if pair.Slot > latestSavedVerifiedRealmSlot {
-			pendingPairs = append(pendingPairs, pair)
-
-			continue
-		}
-
-		vanguardErr = vanguardHashDB.SaveVanguardHeaderHash(pair.Slot, &types.HeaderHash{
-			Status: types.Skipped,
-		})
-
-		// TODO: when more shard will come we will need to maintain this information
-		pandoraErr = pandoraHeaderHashDB.SavePandoraHeaderHash(pair.Slot, &types.HeaderHash{
-			Status: types.Skipped,
-		})
-
-		if nil != vanguardErr || nil != pandoraErr {
-			log.WithField("vanguardErr", vanguardErr).
-				WithField("pandoraErr", pandoraErr).
-				WithField("realmErr", realmErr).
-				Error("Got error during invalidation of pending queue")
-			break
-		}
-	}
-
-	if nil != vanguardErr || nil != pandoraErr {
-		log.WithField("vanguardErr", vanguardErr).
-			WithField("pandoraErr", pandoraErr).
-			Error("Got error during invalidation of possible skipped pairs")
-		errChan <- databaseErrors{
-			vanguardErr: vanguardErr,
-			pandoraErr:  pandoraErr,
-			realmErr:    realmErr,
-		}
-
-		return
-	}
-
-	for _, pair := range pendingPairs {
-		if nil != pair.VanguardHash {
-			vanguardErr = vanguardHashDB.SaveVanguardHeaderHash(pair.Slot, &types.HeaderHash{
-				Status:     types.Skipped,
-				HeaderHash: pair.VanguardHash.HeaderHash,
-			})
-		}
-
-		// TODO: when more shard will come we will need to maintain this information
-		if len(pair.PandoraHashes) > 0 && nil != pair.PandoraHashes[0] {
-			pandoraErr = pandoraHeaderHashDB.SavePandoraHeaderHash(pair.Slot, &types.HeaderHash{
-				Status:     types.Skipped,
-				HeaderHash: pair.PandoraHashes[0].HeaderHash,
-			})
-		}
-	}
-
-	if nil != vanguardErr || nil != pandoraErr {
-		log.WithField("vanguardErr", vanguardErr).
-			WithField("pandoraErr", pandoraErr).
-			Error("Got error during invalidation of pendingPairs")
-		errChan <- databaseErrors{
-			vanguardErr: vanguardErr,
-			pandoraErr:  pandoraErr,
-			realmErr:    realmErr,
-		}
-
-		return
-	}
-
-	var (
-		finalVanguardBatch []*types.HeaderHash
-		finalPandoraBatch  []*types.HeaderHash
-	)
-
-	// At the very end fill all vanguard and pandora nil entries as skipped ones
-	// Do not fetch any higher records
-	finalVanguardBatch, vanguardErr = vanguardHashDB.VanguardHeaderHashes(
-		invalidationStartRealmSlot,
-		invalidationRange,
-	)
-
-	finalPandoraBatch, pandoraErr = pandoraHeaderHashDB.PandoraHeaderHashes(
-		invalidationStartRealmSlot,
-		invalidationRange,
-	)
-
-	if nil != vanguardErr || nil != pandoraErr {
-		log.WithField("vanguardErr", vanguardErr).
-			WithField("pandoraErr", pandoraErr).
-			WithField("realmErr", realmErr).
-			Error("Got error during invalidation of pending queue")
-
-		errChan <- databaseErrors{
-			vanguardErr: vanguardErr,
-			pandoraErr:  pandoraErr,
-			realmErr:    realmErr,
-		}
-
-		return
-	}
-
-	for index, headerHash := range finalVanguardBatch {
-		if nil != headerHash {
-			continue
-		}
-
-		slotToCheck := fromSlot + uint64(index)
-
-		if slotToCheck > latestSavedVerifiedRealmSlot {
-			continue
-		}
-
-		headerHash = &types.HeaderHash{
-			HeaderHash: kv.EmptyHash,
-			Status:     types.Skipped,
-		}
-
-		vanguardErr = vanguardHashDB.SaveVanguardHeaderHash(slotToCheck, headerHash)
-	}
-
-	for index, headerHash := range finalPandoraBatch {
-		if nil != headerHash {
-			continue
-		}
-
-		slotToCheck := fromSlot + uint64(index)
-
-		if slotToCheck > latestSavedVerifiedRealmSlot {
-			continue
-		}
-
-		headerHash = &types.HeaderHash{
-			HeaderHash: kv.EmptyHash,
-			Status:     types.Skipped,
-		}
-		pandoraErr = pandoraHeaderHashDB.SavePandoraHeaderHash(slotToCheck, headerHash)
-	}
-
-	if nil != vanguardErr || nil != pandoraErr {
-		log.WithField("vanguardErr", vanguardErr).
-			WithField("pandoraErr", pandoraErr).
-			WithField("realmErr", realmErr).
-			Error("Got error during invalidation of final Vanguard or Pandora batch")
-
-		errChan <- databaseErrors{
-			vanguardErr: vanguardErr,
-			pandoraErr:  pandoraErr,
-			realmErr:    realmErr,
-		}
-
-		return
-	}
-
-	log.WithField("highestCheckedSlot", latestSavedVerifiedRealmSlot).
-		Info("I have resolved Canonicalize")
 }
 
 func merge(cs ...<-chan *types.HeaderHash) <-chan *types.HeaderHash {
